+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 
 namespace ricaun.Revit.UI.Example.Revit
 {
-    [AppLoader]
+    //[AppLoader]
     //[Obsolete]
     public class App : IExternalApplication
     {
@@ -277,9 +277,6 @@ namespace ricaun.Revit.UI.Example.Revit
                 split.CreatePushButton<Commands.Command, Commands.Availability.AvailableOnAnyDocument>()
                     .SetLargeImage("/UIFrameworkRes;component/ribbon/images/revit.ico");
             }
-
-
-
 
             //Console.WriteLine($">> {split}");
 
@@ -623,6 +620,7 @@ namespace ricaun.Revit.UI.Example.Revit
         {
             ribbonPanelMove?.Remove(true);
             ribbonPanel?.Remove(true);
+
             return Result.Succeeded;
         }