+++ ricaun.Revit.UI.Tests/AppLoaderTests.cs
@@ -1,0 +1,14 @@
+﻿using NUnit.Framework;
+
+namespace ricaun.Revit.UI.Tests
+{
+    public class AppLoaderTests
+    {
+        [Test]
+        public void HasAppLoaderAttribute()
+        {
+            var name = typeof(AppLoaderAttribute).Name;
+            Assert.AreEqual("AppLoaderAttribute", name);
+        }
+    }
+}