+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.6] / 2023-11-08
+### Features
+- Update Remove and QuickAccess
+### Updated
+- Update `QuickAccessToolBarExtension` with internal methods.
+- Update `MoveToRibbonTab` to move to `Modify` tab.
+### Tests
+- Add Tests for `QuickAccessToolBarExtension`
+- Add Tests for `RibbonHelp` and `RibbonItem`
+- Add Tests for `AppLoader`
+- Add Tests for `MoveToRibbonTab`
+
 ## [0.5.5] / 2023-11-03
 ### Features
 - Language Update and Test
@@ -279,6 +291,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.5.6]: ../../compare/0.5.5...0.5.6
 [0.5.5]: ../../compare/0.5.4...0.5.5
 [0.5.4]: ../../compare/0.5.3...0.5.4
 [0.5.3]: ../../compare/0.5.2...0.5.3